Move pollution model initialisation to retrieve_static_boundaries.py
