+++ src/digitaltwin/retrieve_static_boundaries.py
@@ -44,6 +44,8 @@ def main(
     data_to_db.store_geospatial_layers_data_to_db(engine, catchment_area)
     # Store user log information in the database
     data_to_db.user_log_info_to_db(engine, catchment_area)
+    # Read roof_surface file and store in the database
+    data_to_db.save_roof_surface_type_data_to_db(engine)
 
 
 if __name__ == "__main__":
